Address validation feedback and finalize agent formatter package
Agent-Logs-Url: https://github.com/voku/phpstan-agent-format/sessions/5fd0a0dd-f55f-4f6f-8cfc-9da8f5caad67

+++ src/Budget/TokenBudgetReducer.php
@@ -82,7 +82,7 @@ final readonly class TokenBudgetReducer
             $payload[] = $cluster->toArray();
         }
 
-        return (int) max(1, (int) ceil(strlen((string) json_encode($payload, JSON_THROW_ON_ERROR)) / 4));
+        return (int) max(1, ceil(strlen((string) json_encode($payload, JSON_THROW_ON_ERROR)) / 4));
     }
 
     /**
